+++ packages/components/src/components/List/components/Header/components/AllFiltersModal/FilterAccordion.tsx
@@ -2,77 +2,34 @@ import React, { type FC } from "react";
 import AccordionView from "@/views/AccordionView";
 import HeadingView from "@/views/HeadingView";
 import ContentView from "@/views/ContentView";
-import RadioGroupView from "@/views/RadioGroupView";
-import RadioView from "@/views/RadioView";
-import CheckboxGroupView from "@/views/CheckboxGroupView";
-import CheckboxView from "@/views/CheckboxView";
 import type { Filter } from "@/components/List/model/filter/Filter";
-import ButtonView from "@/views/ButtonView";
-import { Flex } from "@/components/Flex";
-import { useLocalizedStringFormatter } from "react-aria";
-import locales from "../../../../locales/*.locale.json";
+import { FilterAccordionRadioGroup } from "@/components/List/components/Header/components/AllFiltersModal/FilterAccordionRadioGroup";
+import { FilterAccordionCheckboxGroup } from "@/components/List/components/Header/components/AllFiltersModal/FilterAccordionCheckboxGroup";
+import { FilterAccordionDateRange } from "@/components/List/components/Header/components/AllFiltersModal/FilterAccordionDateRange";
+import { DateRangeFilter } from "@/components/List/model/filter/DateRangeFilter";
 
 interface Props {
-  filter: Filter<never, never, never>;
+  filter: Filter | DateRangeFilter;
 }
 
 export const FilterAccordion: FC<Props> = (props) => {
   const { filter } = props;
 
-  const activeKeys = filter.values.filter((v) => v.isActive).map((v) => v.id);
-
   const name = filter.name ?? filter.property;
 
-  const stringFormatter = useLocalizedStringFormatter(locales);
+  const content =
+    filter instanceof DateRangeFilter ? (
+      <FilterAccordionDateRange filter={filter} />
+    ) : filter.mode === "one" ? (
+      <FilterAccordionRadioGroup filter={filter} />
+    ) : (
+      <FilterAccordionCheckboxGroup filter={filter} />
+    );
 
   return (
     <AccordionView>
       <HeadingView>{name}</HeadingView>
-      <ContentView>
-        {filter.mode === "one" && (
-          <Flex direction="column" gap="m">
-            <RadioGroupView
-              value={activeKeys[0]}
-              m={[1, 1]}
-              key={activeKeys[0]}
-              aria-label={name}
-            >
-              {filter.values.map((v) => (
-                <RadioView
-                  key={v.id}
-                  value={v.id}
-                  onPress={() => {
-                    if (!v.isActive) {
-                      v.toggle();
-                    }
-                  }}
-                >
-                  {v.render()}
-                </RadioView>
-              ))}
-            </RadioGroupView>
-            {activeKeys.length > 0 && (
-              <ButtonView
-                size="s"
-                color="secondary"
-                variant="soft"
-                onPress={() => filter.clear()}
-              >
-                {stringFormatter.format("list.filters.clearSelection")}
-              </ButtonView>
-            )}
-          </Flex>
-        )}
-        {filter.mode !== "one" && (
-          <CheckboxGroupView value={activeKeys} m={[1, 1]} aria-label={name}>
-            {filter.values.map((v) => (
-              <CheckboxView key={v.id} value={v.id} onPress={() => v.toggle()}>
-                {v.render()}
-              </CheckboxView>
-            ))}
-          </CheckboxGroupView>
-        )}
-      </ContentView>
+      <ContentView>{content}</ContentView>
     </AccordionView>
   );
 };