+++ packages/components/src/components/List/model/filter/types.ts
@@ -4,8 +4,10 @@ import type {
   PropertyValueRenderMethod,
 } from "@/components/List/model/types";
 import type { ItemType } from "@/lib/types/array";
+import type { RangeCalendarProps } from "@/components/Calendar";
+import type { DateRangeFilter } from "@/components/List/model/filter/DateRangeFilter";
 
-export type FilterMode = "all" | "some" | "one";
+export type FilterMode = "all" | "some" | "one" | "dateRange";
 
 export type FilterMatcher<T, P, TMatcherValue> = (
   filterBy: NonNullable<ItemType<TMatcherValue>>,
@@ -24,6 +26,12 @@ export interface FilterShape<T, TProp extends PropertyName<T>, TMatcherValue> {
   priority?: "primary" | "secondary";
   autosave?: boolean;
   manualSave?: boolean;
+  dateRangeOptions?: RangeCalendarProps;
 }
 
 export type FilterUpdatedCallback = (values: unknown[]) => unknown;
+
+export type AnyDateRangeFilter =
+  | DateRangeFilter
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  | DateRangeFilter<any, any>;