+++ packages/remote-elements/src/auto-generated/RemoteRangeCalendarElement.ts
@@ -1,0 +1,128 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import { FlowRemoteElement } from "@/lib/FlowRemoteElement";
+import type { RangeCalendarProps as RemoteRangeCalendarElementProps } from "@mittwald/flow-react-components";
+export type { RangeCalendarProps as RemoteRangeCalendarElementProps } from "@mittwald/flow-react-components";
+
+export class RemoteRangeCalendarElement extends FlowRemoteElement<RemoteRangeCalendarElementProps> {
+  static override get remoteAttributes() {
+    return ["style"];
+  }
+
+  static override get remoteProperties() {
+    return {
+      allowsNonContiguousRanges: {},
+      "aria-describedby": {},
+      "aria-details": {},
+      "aria-label": {},
+      "aria-labelledby": {},
+      autoFocus: {},
+      className: {},
+      createCalendar: {},
+      defaultFocusedValue: {},
+      defaultValue: {},
+      dir: {},
+      firstDayOfWeek: {},
+      focusedValue: {},
+      hidden: {},
+      id: {},
+      inert: {},
+      isDateUnavailable: {},
+      isDisabled: {},
+      isInvalid: {},
+      isReadOnly: {},
+      lang: {},
+      maxValue: {},
+      minValue: {},
+      pageBehavior: {},
+      selectionAlignment: {},
+      slot: {},
+      translate: {},
+      value: {},
+      visibleDuration: {},
+    };
+  }
+
+  static override get remoteEvents() {
+    return {
+      animationEnd: {},
+      animationEndCapture: {},
+      animationIteration: {},
+      animationIterationCapture: {},
+      animationStart: {},
+      animationStartCapture: {},
+      auxClick: {},
+      auxClickCapture: {},
+      change: {},
+      click: {},
+      clickCapture: {},
+      contextMenu: {},
+      contextMenuCapture: {},
+      doubleClick: {},
+      doubleClickCapture: {},
+      focusChange: {},
+      gotPointerCapture: {},
+      gotPointerCaptureCapture: {},
+      lostPointerCapture: {},
+      lostPointerCaptureCapture: {},
+      mouseDown: {},
+      mouseDownCapture: {},
+      mouseEnter: {},
+      mouseLeave: {},
+      mouseMove: {},
+      mouseMoveCapture: {},
+      mouseOut: {},
+      mouseOutCapture: {},
+      mouseOver: {},
+      mouseOverCapture: {},
+      mouseUp: {},
+      mouseUpCapture: {},
+      pointerCancel: {},
+      pointerCancelCapture: {},
+      pointerDown: {},
+      pointerDownCapture: {},
+      pointerEnter: {},
+      pointerLeave: {},
+      pointerMove: {},
+      pointerMoveCapture: {},
+      pointerOut: {},
+      pointerOutCapture: {},
+      pointerOver: {},
+      pointerOverCapture: {},
+      pointerUp: {},
+      pointerUpCapture: {},
+      scroll: {},
+      scrollCapture: {},
+      touchCancel: {},
+      touchCancelCapture: {},
+      touchEnd: {},
+      touchEndCapture: {},
+      touchMove: {},
+      touchMoveCapture: {},
+      touchStart: {},
+      touchStartCapture: {},
+      transitionCancel: {},
+      transitionCancelCapture: {},
+      transitionEnd: {},
+      transitionEndCapture: {},
+      transitionRun: {},
+      transitionRunCapture: {},
+      transitionStart: {},
+      transitionStartCapture: {},
+      wheel: {},
+      wheelCapture: {},
+    };
+  }
+
+  static override get remoteSlots() {
+    return [];
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    "flr-range-calendar": InstanceType<typeof RemoteRangeCalendarElement>;
+  }
+}
+
+customElements.define("flr-range-calendar", RemoteRangeCalendarElement);