+++ packages/remote-elements/src/auto-generated/index.ts
@@ -93,6 +93,7 @@ export * from "./RemoteProgressBarElement";
 export * from "./RemoteRadioElement";
 export * from "./RemoteRadioButtonElement";
 export * from "./RemoteRadioGroupElement";
+export * from "./RemoteRangeCalendarElement";
 export * from "./RemoteRatingElement";
 export * from "./RemoteSearchFieldElement";
 export * from "./RemoteSectionElement";